write types for app module, delete actions for data language
