--- src/store/app/actionTypes.ts
@@ -1,9 +1,0 @@
-/**
- * Set another language for displaying data from database
- */
-export const CHANGE_DATA_LANG = 'CHANGE_DATA_LANG';
-
-/**
- * Set another language for user interface
- */
-export const CHANGE_INTERFACE_LANG = 'CHANGE_INTERFACE_LANG';
