+++ src/store/modules/app/actionTypes.ts
@@ -1,0 +1,4 @@
+/**
+ * Set another language for user interface
+ */
+export const CHANGE_INTERFACE_LANG = 'CHANGE_INTERFACE_LANG';
